+++ CHANGELOG.md
@@ -7,6 +7,10 @@ Full documentation for amd_smi_lib is available at [https://rocm.docs.amd.com/pr
 
 ### Added
 
+- **Added dynamic virtualization mode detection**.  
+  - Added new C and Python API `amdsmi_get_gpu_virtualization_mode_info`
+  - Added new C and Python enum `amdsmi_virtualization_mode_t`
+
 - **Added TVIOL_ACTIVE to `amd-smi monitor`**.  
 Added temperature violation active or not status to `amd-smi monitor`. TVIOL_ACTIVE will be displayed as below:
  - True if active
@@ -135,14 +139,23 @@ GPU: 0
 
 ### Changed
 
+- **Corrected CLI CPU argument name**.  
+  - `--cpu-pwr-svi-telemtry-rails` to `--cpu-pwr-svi-telemetry-rails`
+
 - **Added amdgpu driver version and amd_hsmp driver version to `amd-smi version` command**.  
   - The `amd-smi version` command can now also display the amdgpu driver version using the `-g` flag.
   - The amd_hsmp driver version can also be displayed using the `-c` flag.
   - The new default for the `version` command is to display all the version information, including both amdgpu and amd_hsmp driver versions.
 
 ```shell
+amd-smi version
+AMDSMI Tool: 24.7.1+b446d6c-dirty | AMDSMI Library version: 24.7.2.0 | ROCm version: N/A | amdgpu version: 6.10.10 | amd_hsmp version: 2.2
+
 amd-smi version -g
-AMDSMI Tool: 24.7.1+6fa991c-dirty | AMDSMI Library version: 24.7.2.0 | ROCm version: N/A | amdgpu version: 6.12.2
+AMDSMI Tool: 24.7.1+b446d6c-dirty | AMDSMI Library version: 24.7.2.0 | ROCm version: N/A | amdgpu version: 6.10.10
+
+amd-smi version -c
+AMDSMI Tool: 24.7.1+b446d6c-dirty | AMDSMI Library version: 24.7.2.0 | ROCm version: N/A | amd_hsmp version: 2.2
 ```
 
 - **All `amd-smi set` and `amd-smi reset` options are now mutually exclusive**.